+++ src/pages/admin/AdminEvents.jsx
@@ -109,9 +109,11 @@ export default function AdminEvents() {
       gap: "0.5rem",
     },
     dirty: {
+      display: "flex",
       textTransform: 'uppercase',
       color: "red",
-      marginRight: "1rem"
+      marginRight: "1rem",
+      alignItems: "center",
     }
   }
 
@@ -467,7 +469,18 @@ export default function AdminEvents() {
 
         </Stack>
       </Box>
-      <Box style={style.actions}>
+
+      <Box sx={{
+          position: 'sticky',
+          bottom: 0,
+          backgroundColor: 'background.paper',
+          padding: 2,
+          display: 'flex',
+          justifyContent: 'space-between',
+          alignItems: 'center',
+          zIndex: 1,
+          borderTop: `1px solid ${theme.palette.divider}`,
+        }}>
         <Box style={{ display: "flex", gap: "0.5rem" }}>
           { isUpdate && <IconButton onClick={() => handleDelete(updateId)} variant="outlined" color="warning"><DeleteIcon /></IconButton> }
           { isUpdate && <Button onClick={handleDuplicate} variant='outlined'>Duplicate</Button> }