+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Admin:
   - Add: Image upload now automatically uploads after a file is selected
   - Add: You can now delete an image (and you are challenged to confirm)
+  - Add: Admin screens now always have their bottom action bars visible
 
 App:
   - Add: Upgrade to React 19